Divide scheduler module, add .gitignore file
